1.添加组件化框架 2.添加屏幕适配工具类ScreenUtil

+++ app/src/main/java/com/example/administrator/androidtest/Common/Util/ResourceUtil.java
@@ -43,7 +43,7 @@ public class ResourceUtil extends AppInitUtil {
     /**
      * 长度资源id转px(dp -> px)返回整数
      */
-    public static float getDimensionPixelSize(int id) {
+    public static int getDimensionPixelSize(int id) {
         return getResources().getDimensionPixelSize(id);
     }
